content/canvas/test/test_2d.drawImage.negativedest.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.drawImage.negativedest</title>
<script src="/MochiKit/MochiKit.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>
var _deferred = false;

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ctx.fillStyle = '#0f0';
ctx.fillRect(0, 0, 100, 50);
var _thrown = undefined; try {
  ctx.drawImage(document.getElementById('red.png'), 10, 10, 1, 1, 0, 50, -100, 50);
} catch (e) { _thrown = e }; ok(_thrown && _thrown.code == DOMException.INDEX_SIZE_ERR, "should throw INDEX_SIZE_ERR");
var _thrown = undefined; try {
  ctx.drawImage(document.getElementById('red.png'), 10, 10, 1, 1, 0, 50, 100, -50);
} catch (e) { _thrown = e }; ok(_thrown && _thrown.code == DOMException.INDEX_SIZE_ERR, "should throw INDEX_SIZE_ERR");

if (!_deferred) SimpleTest.finish();
});
</script>
<img src="image_red.png" id="red.png" class="resource">
